Refactor Environment class (#70)
* Add error rollback decorator and context manager

* Move Environment class declaration

* Refactor Environment

- allow Extractor plugins (instead of only SourceExtractor)
- make_converter always returns a callable
- add make_transform
