+++ datumaro/util/os_util.py
@@ -1,9 +1,12 @@
 
-# Copyright (C) 2020 Intel Corporation
+# Copyright (C) 2020 - 2021 Intel Corporation
 #
 # SPDX-License-Identifier: MIT
 
+import importlib
+import os.path as osp
 import subprocess
+import sys
 
 
 def check_instruction_set(instruction):
@@ -15,3 +18,17 @@ def check_instruction_set(instruction):
             'lscpu | grep -o "%s" | head -1' % instruction,
             shell=True).decode('utf-8') # nosec
     )
+
+def import_foreign_module(name, path, package=None):
+    module = None
+    default_path = sys.path.copy()
+    try:
+        sys.path = [ osp.abspath(path), ] + default_path
+        sys.modules.pop(name, None) # remove from cache
+        module = importlib.import_module(name, package=package)
+        sys.modules.pop(name) # remove from cache
+    except Exception:
+        raise
+    finally:
+        sys.path = default_path
+    return module
